requirements.txt
asgiref==3.3.4
Django==3.2.2
django-cors-headers==3.7.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.9.2
PyJWT==2.1.0
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.7.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
